Move components to src/lib/components
We're doing this to avoid needing to use relative paths when including
components.
