+++ src/lib/components/index.ts
@@ -1,0 +1,3 @@
+export { default as Message } from './Message.svelte';
+export { default as Spinner } from './Spinner.svelte';
+export { default as Warning } from './Warning.svelte';
